Adds the ability to create asp.net containers through a restful interface.
On Windows, we are currently defining a container to be an IIS site
that contains exactly one application, and has a one to one correspondence
with an application pool. A container can be created by posting to the
"/api/Containers/" endpoint. The container's id will be contained in the
response.

[Finishes #82044594]