Fix port mapping
Previously containerizer assumed that the port mapping is always going
to be host => 8080. With the introduction of sshd in a recent diego
change, we are getting two calls to NetIn and returning port mapping
from `[{hostport1: 8080},{hostport2: 8080}]` which is confusing the
router. This change returns the right port mapping (which was a todo for
a while anyway)

[#94444368]