Fix killAllProcs helper
We were panicking when the parent process was nil, which made it so that
if the containerizer/garden was leftover from an exited test run, it
would blow up. This considerably simplifies it while I believe still
covering the bases:

1. A nil parent is assumed to be orphaned by a prior test run
1. A parent with "test" in the name is assumed to be orphaned by an
earlier test in the same test run

Might still be annoying if you are running a containerizer/garden in a
separate window, but alas.