Retries deleting the container for 5 times with 1 second delay
Submodule Containerizer/IronFrame 7cfa701..75cd720:
  > Do not remove the container unless the user and directory were removed successfully

[#99822778]

This hopefully resolves the issue where deleting a container fails when the
files are being held open by Containerizer itself. This can happen when a
StreamIn is happening at the same time as a deletion.

The ideal solution would probably be to interrupt the StreamIn action before
trying to delete, but this was more complicated to accomplish.

We aren't sure if our 5x1s retry logic is the right interval/amount of retries,
but it is a best guess for what we understand about the problem right now.

Signed-off-by: John Shahid <jshahid@pivotal.io>