Streamin file to a different destination
Beacuse the quota was applied before the inital stream in,
and the second assterted StreamIn would replace the first file
which was above the set quota, the test would fail from time to time.

Also remove the assertion the the StreamIn fail if the quota is excedded,
because it doesn't work everytime. The disk quota is still enforced, even
though the StreamIn failed.