Prevent container destroy & container index concurrency
This prevents a race condition where a container would be in the process
of being destroyed and a separate request would try to read its
properties.json and fail.

This issue still can occur if the container fails to be deleted but its
properties.json is deleted, so we additionally added a try/catch for
DirectoryNotFoundException.

[#104436838]

Signed-off-by: Ben Moss <bmoss@pivotal.io>